+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
-
+- Added `_api` property in `Connection` class.
+  
 ## v0.0.1
 - Initial version.